refactor(components/atom/button): add import file extensions
