+++ components/atom/button/src/buttonSpinnerIcon/config.js
@@ -1,3 +1,3 @@
-import {CLASS} from '../config'
+import {CLASS} from '../config.js'
 
 export const BASE_CLASSNAME = `${CLASS}SpinnerIcon`
