+++ components/atom/button/src/config.js
@@ -1,3 +1,5 @@
+import {cloneElement} from 'react'
+
 /**
  * Base class for the component
  */
@@ -133,3 +135,120 @@ export const TYPES_CONVERSION = {
 
 export const createClasses = (array, sufix = '') =>
   array.reduce((res, key) => ({...res, [key]: `${CLASS}--${key}${sufix}`}), {})
+
+export const CLASSES = createClasses([
+  ...COLORS,
+  ...Object.values(DESIGNS),
+  ...Object.values(ALIGNMENT),
+  ...MODIFIERS,
+  ...Object.values(SIZES),
+  'empty'
+])
+
+/**
+ * Get props cleaning out AtomButton own props
+ * @param  {Object} props
+ * @return {Object}
+ */
+export const cleanProps = props => {
+  const newProps = {...props}
+  OWN_PROPS.forEach(key => delete newProps[key])
+  return newProps
+}
+
+/**
+ * Get modifiers to apply according to props
+ * @param  {Object} props
+ * @return {Array<String>}
+ */
+export const getModifiers = props => {
+  return Object.keys(props).filter(
+    name => props[name] && MODIFIERS.includes(name)
+  )
+}
+
+export function deprecated(
+  validator,
+  callback = (props, propName, componentName) => {
+    const deprecatedMessage = `The prop ${'\x1b[32m'}${propName}${'\u001b[39m'} is DEPRECATED on ${'\x1b[32m'}${componentName}${'\u001b[39m'}.`
+    console.warn(deprecatedMessage) // eslint-disable-line
+  }
+) {
+  return function deprecated(props, propName, componentName, ...rest) {
+    if (props[propName] != null && process.env.NODE_ENV === 'development') {
+      callback(props, propName, componentName, ...rest)
+    }
+    return validator(props, propName, componentName, ...rest)
+  }
+}
+
+export const typeConversion = ({type, design, color, link, href, ...other}) => {
+  const result = {
+    design,
+    color,
+    link,
+    href,
+    ...other
+  }
+  switch (type) {
+    case 'primary':
+      result.color = color || 'primary'
+      result.design = design || (link || href ? DESIGNS.LINK : DESIGNS.SOLID)
+      break
+    case 'accent':
+      result.color = color || 'accent'
+      result.design = design || (link || href ? DESIGNS.LINK : DESIGNS.SOLID)
+      break
+    case 'secondary':
+      result.color = color || 'primary'
+      result.design = design || (link || href ? DESIGNS.LINK : DESIGNS.OUTLINE)
+      break
+    case 'tertiary':
+      result.color = color || 'primary'
+      result.design = design || (link || href ? DESIGNS.LINK : DESIGNS.FLAT)
+      break
+    default:
+      result.type = type
+      result.color = color || 'primary'
+      break
+  }
+  return result
+}
+
+export const getPropsWithDefaultValues = ({
+  type,
+  design,
+  color,
+  alignment,
+  link,
+  href,
+  ...other
+}) => ({
+  ...other,
+  link,
+  type,
+  design: design || (link || href ? DESIGNS.LINK : DESIGNS.SOLID),
+  color: color || 'colors',
+  alignment: alignment || ALIGNMENT.CENTER
+})
+
+/**
+ * Detect if element is an AtomIcon to force correct size
+ * @param {React.ReactElement} icon
+ */
+export const isAtomIcon = icon =>
+  icon?.type?.displayName === ATOM_ICON_DISPLAY_NAME
+
+/**
+ * Prepare the AtomIcon element to use the correct size
+ * @param {React.ReactElement} atomIconElement
+ * @param {object} options
+ * @param {string} options.size Size of the button to grab the correct icon size
+ */
+export const prepareAtomIcon = (atomIconElement, {size}) => {
+  const atomIconSize = ATOM_ICON_SIZES_MAPPER[size]
+  return cloneElement(atomIconElement, {
+    color: undefined,
+    size: atomIconSize
+  })
+}
