+++ components/atom/button/test/index.test.js
@@ -9,11 +9,10 @@ import chaiDOM from 'chai-dom'
 import sinon from 'sinon'
 import {fireEvent} from '@testing-library/react'
 
-import * as pkg from '../src'
+import * as pkg from '../src/index.js'
+import {createClasses} from '../src/config.js'
 
 import json from '../package.json'
-
-import {createClasses} from '../src/config'
 
 chai.use(chaiDOM)
